List page newdesignintegration

+++ src/scripts/list.js
@@ -1,6 +1,5 @@
-//Store the data into localStorage before staring all the things.
-let journalList = getJournalList();
-
+// Instantiate list of entries.
+let journalList;
 document.addEventListener("DOMContentLoaded", init);
 
 /**
@@ -8,23 +7,113 @@ document.addEventListener("DOMContentLoaded", init);
  * @typedef {Object} JournalEntry
  * @property {number} timestamp - unique identifier and time it was created
  * @property {string} title - title of the journal entry
- * @property {Array.string} tags - list of tags
+ * @property {string[]} tags - list of tags
  * @property {Object} delta - Quill delta containing text operations
  */
-
 let quill;
+
+let sortDirection = {
+  name: true,
+  timestamp: true,
+};
 
 /**
  * Called on page load.
  */
 function init() {
+  // Store the data into localStorage before starting all the things.
+  journalList = getJournalList();
   const newJournalButton = document.querySelector(".new-journal-button");
+  const filterButton = document.querySelector(".filter-button");
   displayList(journalList);
+
   setUpSearch();
 
   newJournalButton.addEventListener("click", function () {
     editJournal();
   });
+
+  // Animation for the filter dropdown
+  filterButton.addEventListener("click", function () {
+    const filterHeader = document.querySelector('.filter-container');
+    const entryHeader = document.querySelector('.entry-header');
+    if (filterHeader.classList.contains('show')) {
+      filterHeader.classList.remove('show');
+      entryHeader.style.marginTop = '75px'; // adjust based on filterHeader height
+      setTimeout(function () {
+        filterHeader.style.display = 'none';
+      }, 500);
+    } else {
+      filterHeader.style.display = 'grid';
+      setTimeout(() => {
+        filterHeader.classList.add('show');
+        entryHeader.style.marginTop = '105px'; // adjust based on filterHeader height
+      }, 0);
+    }
+  });
+
+}
+
+document.getElementById("sort-name").addEventListener("click", () => {
+  sortByCategory("name");
+});
+document.getElementById("sort-timestamp").addEventListener("click", () => {
+  sortByCategory("timestamp");
+});
+
+/**
+ * Sorts the journal list by the specified category.
+ *
+ * @param {string} category - The category to sort by. This can be "name" or "timestamp".
+ * If "name" is specified, the journal list is sorted alphabetically by the title of the journal entries.
+ * If "timestamp" is specified, the journal list is sorted chronologically by the timestamp of the journal entries.
+ * The sort direction (ascending or descending) is toggled each time the function is called with the same category.
+ */
+function sortByCategory(category) {
+  if (category === "name") {
+    journalList.sort((a, b) => {
+      if (sortDirection.name) {
+        return a.title.localeCompare(b.title);
+      } else {
+        return b.title.localeCompare(a.title);
+      }
+    });
+    sortDirection.name = !sortDirection.name;
+  } else if (category === "timestamp") {
+    journalList.sort((a, b) => {
+      if (sortDirection.editTime) {
+        return a.editTime - b.editTime;
+      } else {
+        return b.editTime - a.editTime;
+      }
+    });
+    sortDirection.editTime = !sortDirection.editTime;
+  }
+  updateSortArrows(category);
+  displayList(journalList);
+}
+
+/**
+ * Updates the direction of the sorting arrows based on the sorting category and direction.
+ *
+ * @param {string} category - The category of sorting. This can be "name" or "timestamp".
+ * If "name" is specified, the name sort arrow's direction is updated based on the sort direction.
+ * If "timestamp" is specified, the timestamp sort arrow's direction is updated based on the sort direction.
+ * The other arrow is reset to its default direction.
+ */
+function updateSortArrows(category) {
+  const nameSortArrow = document.getElementById("sort-name");
+  const timestampSortArrow = document.getElementById("sort-timestamp");
+
+  if (category === "name") {
+    nameSortArrow.innerHTML = sortDirection.name ? "&#9650;" : "&#9660;";
+    timestampSortArrow.innerHTML = "&#9650;"; // Reset the other arrow
+  } else if (category === "timestamp") {
+    timestampSortArrow.innerHTML = sortDirection.editTime
+      ? "&#9650;"
+      : "&#9660;";
+    nameSortArrow.innerHTML = "&#9650;"; // Reset the other arrow
+  }
 }
 
 /**
@@ -50,23 +139,17 @@ function createListItem(item) {
   const itemList = document.getElementById("item-list");
   const listItem = document.createElement("li");
 
+  //Create title container
   const title = document.createElement("div");
+  title.setAttribute("id", "entry-title");
   title.textContent = item.title;
+  title.className = "title";
   listItem.appendChild(title);
 
-  const details = document.createElement("div");
-  details.style.fontSize = "small";
-
-  let timestamp = parseInt(item.timestamp);
-  const timestampText = document.createElement("div");
-  timestampText.textContent = `Timestamp: ${new Date(
-    timestamp,
-  ).toLocaleString()}`;
-  details.appendChild(timestampText);
-
-  // Generate tags
+  //Generate tags and create tag container
   const tagsContainer = document.createElement("div");
-  tagsContainer.textContent = "Tags: ";
+  tagsContainer.setAttribute("id", "entry-tags");
+  tagsContainer.textContent = "";
 
   item.tags.forEach((tag) => {
     const tagElement = document.createElement("span");
@@ -79,22 +162,49 @@ function createListItem(item) {
     tagsContainer.appendChild(document.createTextNode(" ")); // Add space between tags
   });
 
-  details.appendChild(tagsContainer);
-  listItem.appendChild(details);
+  listItem.appendChild(tagsContainer);
 
+  //Create timestamp container
+  let timestamp = parseInt(item.timestamp);
+  const timestampText = document.createElement("div");
+
+  timestampText.setAttribute("id", "entry-timestamp");
+
+  let date = new Date(item.editTime);
+
+  let options = {
+    year: 'numeric',
+    month: '2-digit',
+    day: '2-digit',
+    hour: '2-digit',
+    minute: '2-digit',
+    hour12: false // Use 24-hour time
+  };
+
+  // Display the formatted date
+  timestampText.textContent = date.toLocaleString('en-GB', options).replace(',', '');
+
+  listItem.appendChild(timestampText);
+
+  //Create delete button
+  const deleteButtonContainer = document.createElement("div");
   const deleteButton = document.createElement("button");
   deleteButton.textContent = "Delete";
   deleteButton.className = "delete-button";
-  deleteButton.style.display = "none";
+  deleteButtonContainer.id = "delete-container"
+  deleteButtonContainer.appendChild(deleteButton);
 
   // EventListener: When clicking delete, delete from page and LocalStorage
   deleteButton.onclick = (event) => {
+    if (window.confirm(`Are you sure you would like to delete the "${item.title}"?`)) {
+      event.stopPropagation();
+      listItem.remove();
+      deleteJournal(timestamp);
+    }
     event.stopPropagation();
-    listItem.remove();
-    deleteJournal(timestamp);
   };
 
-  listItem.appendChild(deleteButton);
+  listItem.appendChild(deleteButtonContainer);
   listItem.onmouseover = () => {
     deleteButton.style.display = "inline";
   };
@@ -127,7 +237,7 @@ function getJournalList() {
  * @returns {JournalEntry|undefined} - journal entry or undefined if not found
  */
 function getJournalByTimestamp(timestamp) {
-  journal = journalList.find((entry) => entry.timestamp == timestamp);
+  let journal = journalList.find((entry) => entry.timestamp == timestamp);
   if (journal === undefined) {
     console.error(`Error: No journal entry found with timestamp ${timestamp}`);
     return undefined;
@@ -152,14 +262,31 @@ function saveJournalList(journalList) {
 }
 
 /**
+ * Checks if a title is valid.
+ * @param title {string} - title of the journal entry
+ * @returns {boolean} - true if title is valid, false otherwise
+ */
+function isTitleValid(title) {
+  return title.trim().length > 0;
+}
+
+/**
+ * Default title for a journal entry.
+ * @type {string}
+ */
+const DEFAULT_TITLE = "Untitled";
+
+/**
  * Opens a modal to edit a journal entry.
  * @param id {number} - unique identifier and time it was created
  */
 function editJournal(id) {
   const modal = document.getElementById("journalModal");
-  const closeModal = document.getElementById("closeModal");
-  const saveJournal = document.getElementById("saveJournal");
+  /** @type {HTMLButtonElement} */
+  const saveJournal = document.getElementById("closeModal");
+  /** @type {HTMLInputElement} */
   const titleBar = document.getElementById("journalTitle");
+  /** @type {HTMLDivElement} */
   const itemList = document.getElementById("item-list");
 
   modal.style.display = "block";
@@ -168,32 +295,17 @@ function editJournal(id) {
     quill = new Quill("#editor", { theme: "snow" });
   }
 
-  closeModal.addEventListener("click", function () {
-    modal.style.display = "none";
-    displayList(journalList);
-  });
 
-  window.addEventListener("click", function (event) {
-    if (event.target == modal) {
-      modal.style.display = "none";
-    }
-  });
-
-  saveJournal.addEventListener("click", function () {
-    const journalContent = quill.root.innerHTML;
-    console.log(journalContent);
-    modal.style.display = "none";
-    itemList.innerHTML = "";
-    displayList(journalList);
-  });
-
+  let noteObject;
   if (id === undefined) {
     id = new Date().getTime();
-    let noteObject = {
+    noteObject = {
       timestamp: id,
-      title: "",
+      editTime: id,
+      title: DEFAULT_TITLE,
       tags: [],
-      delta: undefined,
+      //delta: undefined,
+      delta: { ops: [] },
     };
 
     quill.setText("\n");
@@ -203,29 +315,102 @@ function editJournal(id) {
 
   noteObject = getJournalByTimestamp(id);
 
-  quill.setContents(noteObject.delta);
+  let contentScreenShot = noteObject.delta;
+
+  quill.setContents(contentScreenShot);
   titleBar.value = noteObject.title;
 
-  quill.on("text-change", () => {
-    const newDelta = quill.getContents();
-    noteObject.delta = newDelta;
-    saveJournalList(journalList);
+  quill.on("text-change", quillUpdateTextHandler);
+
+  titleBar.addEventListener("input", updateTitleHandler);
+
+
+  // Cancel changes and revert notebook
+  const cancelButton = document.getElementById('cancelModal');
+  cancelButton.addEventListener('click', function () {
+    noteObject.delta = contentScreenShot;
+
+    if (contentScreenShot.ops == [] && !isTitleValid(titleBar.value)){
+      deleteJournal(noteObject.timestamp);
+    }
+    else if (!isTitleValid(titleBar.value)) {
+      cancelButton.disabled = true;
+    }
+    else {
+      cancelButton.disabled = false;
+    }
+
+    modal.style.display = "none";
+    itemList.innerHTML = "";
+    displayList(journalList);
+    removeJournalEventListeners();
   });
 
-  titleBar.addEventListener("input", () => {
+
+  saveJournal.addEventListener(
+    "click",
+    function () {
+      updateTitleHandler();
+      quillUpdateTextHandler();
+      modal.style.display = "none";
+      itemList.innerHTML = "";
+      displayList(journalList);
+      removeJournalEventListeners();
+    },
+    { once: true },
+  );
+
+
+  /**
+   * Updates journal entry title with current contents in the title input bar.
+   */
+  function updateTitleHandler() {
     let title = titleBar.value;
     noteObject.title = title;
-    saveJournalList(journalList);
-  });
+
+
+    if (isTitleValid(title)) {
+      // don't save if title is empty
+      saveJournalList(journalList);
+
+      saveJournal.disabled = false;
+    } else {
+      saveJournal.disabled = true;
+      saveJournal.title = "Title cannot be empty";
+    }
+  }
+
+  /**
+   * Updates journal entry with current contents of the Quill editor.
+   */
+  function quillUpdateTextHandler() {
+    const newDelta = quill.getContents();
+    noteObject.delta = newDelta;
+
+    if (isTitleValid(noteObject.title)) {
+      // don't save if title is empty
+      saveJournalList(journalList);
+    }
+
+    noteObject.editTime = new Date().getTime();
+  }
+
+  /**
+   * Removes event listeners on input fields for the current journal.
+   */
+  function removeJournalEventListeners() {
+    titleBar.removeEventListener("input", updateTitleHandler);
+    quill.off("text-change", quillUpdateTextHandler);
+  }
 }
 
 /**
  * Searches all journal entries for a string only if the entries include all the specified tags and is within the time period filter.
  * @param {string} query - exact string to search for
- * @param {Array.string} tags - list of exact tags to include
+ * @param {string[]} tags - list of exact tags to include
  * @param {string} startDate - start date formatted yyyy-mm-dd
  * @param {string} endDate - end date formatted yyyy-mm-dd
- * @returns {JournalEntry[]} matching entries
+ * @returns matching entries
  */
 function searchJournal(query, tags, startDate, endDate) {
   let filteredList = journalList;
@@ -255,9 +440,9 @@ function searchJournal(query, tags, startDate, endDate) {
 
 /**
  * Searches a list of entries for a case-insensitive string.
- * @param {JournalEntry[]} list - list of entries
+ * @param {Array.Object} list - list of entries
  * @param {string} query - exact string to search for
- * @returns {JournalEntry[]} matching entries
+ * @returns matching entries
  */
 function getMatchingEntries(list, query) {
   query = query.toLowerCase();
@@ -279,9 +464,13 @@ function getMatchingEntries(list, query) {
 /**
  * Extracts all the text in a Quill delta.
  * @param {Object} delta - Quill delta containing text operations
- * @returns {string} all the text in a Quill delta
+ * @returns all the text in a Quill delta
  */
 function getTextFromDelta(delta) {
+  if (!delta || !delta.ops) {
+    return '';
+  }
+  
   let text = "";
   delta.ops.forEach((op) => {
     text += op.insert;
@@ -292,7 +481,7 @@ function getTextFromDelta(delta) {
 /**
  * Parses a string of comma-separated tags into an array.
  * @param {string} tagsString - string of comma-separated tags
- * @returns {string[]} array of tags
+ * @returns array of tags
  */
 function parseTags(tagsString) {
   return tagsString.split(",").filter((tag) => tag.length > 0);
@@ -325,3 +514,5 @@ function setUpSearch() {
     };
   });
 }
+
+export { getTextFromDelta, getMatchingEntries };
